+++ nmd_assembly.h
@@ -7728,7 +7728,7 @@ NMD_ASSEMBLY_API void _nmd_append_relative_address16_32(_nmd_string_info* const 
 	if (si->runtime_address == NMD_X86_INVALID_RUNTIME_ADDRESS)
 	{
 		/* *si->buffer++ = '$'; */
-		_nmd_append_signed_number(si, (int64_t)((int32_t)(si->instruction->immediate) + (int32_t)(si->instruction->length)), true);
+		_nmd_append_signed_number(si, (int64_t)(si->instruction->immediate + si->instruction->length), true);
 	}
 	else
     {
@@ -7780,7 +7780,7 @@ NMD_ASSEMBLY_API void _nmd_append_modrm16_upper(_nmd_string_info* const si)
 			_nmd_append_number(si, si->instruction->displacement);
 		else
 		{
-			const bool is_negative = si->instruction->displacement & (1 << (si->instruction->disp_mask * 8 - 1));
+			const bool is_negative = si->instruction->displacement & (1U << (si->instruction->disp_mask * 8 - 1));
 			if (*(si->buffer - 1) != '[')
 				*si->buffer++ = is_negative ? '-' : '+';
